Move pools location (adjusted interface) (#8935)
* Move pools location (adjusted interface)

* Adjust type import
