+++ packages/page-staking/src/types.ts
@@ -2,9 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import type { Inflation } from '@polkadot/react-hooks/types';
-import type { u32 } from '@polkadot/types';
 import type { AccountId, Balance, BlockNumber, EraIndex, Exposure, Hash, SessionIndex, ValidatorPrefs, ValidatorPrefsTo196 } from '@polkadot/types/interfaces';
-import type { PalletNominationPoolsPoolMember } from '@polkadot/types/lookup';
 import type { BN } from '@polkadot/util';
 
 export type Nominators = Record<string, string[]>;
@@ -96,18 +94,4 @@ export interface SortedTargets {
   validators?: ValidatorInfo[];
   validatorIds?: string[];
   waitingIds?: string[];
-}
-
-export interface PoolAccounts {
-  rewardId: string;
-  stashId: string;
-}
-
-export interface OwnPoolBase {
-  members: Record<string, PalletNominationPoolsPoolMember>;
-  poolId: u32;
-}
-
-export interface OwnPool extends OwnPoolBase, PoolAccounts {
-  // nothing additional, only combined
 }
