+++ packages/page-staking2/src/Pools/Join.tsx
