+++ packages/page-staking2/src/Pools/index.tsx
@@ -1,12 +1,13 @@
 // Copyright 2017-2023 @polkadot/app-staking authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
-import type { OwnPool } from '../types';
+import type { OwnPool } from './types';
 
 import React from 'react';
 
 import Pools from './Pools';
 import Summary from './Summary';
+import useOwnPools from './useOwnPools';
 import useParams from './useParams';
 import usePoolIds from './usePoolIds';
 
@@ -15,8 +16,9 @@ interface Props {
   ownPools?: OwnPool[];
 }
 
-function NominationPools ({ className, ownPools }: Props): React.ReactElement<Props> {
+function NominationPools ({ className, ownPools: ownPoolProp }: Props): React.ReactElement<Props> {
   const ids = usePoolIds();
+  const ownPools = useOwnPools();
   const params = useParams();
 
   return (
@@ -27,7 +29,7 @@ function NominationPools ({ className, ownPools }: Props): React.ReactElement<Pr
       />
       <Pools
         ids={ids}
-        ownPools={ownPools}
+        ownPools={ownPools || ownPoolProp}
         params={params}
       />
     </div>
