+++ packages/page-staking2/src/Pools/types.ts
@@ -1,9 +1,23 @@
 // Copyright 2017-2023 @polkadot/app-staking authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
+import type { u32 } from '@polkadot/types';
 import type { PalletNominationPoolsBondedPoolInner, PalletNominationPoolsPoolMember, PalletNominationPoolsRewardPool } from '@polkadot/types/lookup';
 import type { BN } from '@polkadot/util';
-import type { PoolAccounts } from '../types';
+
+export interface PoolAccounts {
+  rewardId: string;
+  stashId: string;
+}
+
+export interface OwnPoolBase {
+  members: Record<string, PalletNominationPoolsPoolMember>;
+  poolId: u32;
+}
+
+export interface OwnPool extends OwnPoolBase, PoolAccounts {
+  // nothing additional, only combined
+}
 
 export interface Params {
   lastPoolId: BN;
