+++ packages/page-staking2/src/Pools/useAmountError.ts
