+++ packages/page-staking2/src/index.tsx
@@ -8,16 +8,33 @@ import { Route, Switch } from 'react-router';
 
 import { Tabs } from '@polkadot/react-components';
 import { useApi, useFavorites } from '@polkadot/react-hooks';
+import { isFunction } from '@polkadot/util';
 
 import { STORE_FAVS_BASE } from './constants';
+import Pools from './Pools';
 import { useTranslation } from './translate';
 import { clearCache } from './useCache';
 import useSessionInfo from './useSessionInfo';
 import Validators from './Validators';
 
+function createPathRef (basePath: string): Record<string, string | string[]> {
+  return {
+    bags: `${basePath}/bags`,
+    payout: `${basePath}/payout`,
+    pools: `${basePath}/pools`,
+    query: [
+      `${basePath}/query/:value`,
+      `${basePath}/query`
+    ],
+    slashes: `${basePath}/slashes`,
+    targets: `${basePath}/targets`
+  };
+}
+
 function StakingApp ({ basePath }: Props): React.ReactElement<Props> {
   const { t } = useTranslation();
   const { api } = useApi();
+  const pathRef = useRef(createPathRef(basePath));
 
   // on unmount anything else, ensure that for the next round we
   // are starting with a fresh cache (there could be large delays)
@@ -41,6 +58,10 @@ function StakingApp ({ basePath }: Props): React.ReactElement<Props> {
       isRoot: true,
       name: 'sign',
       text: t<string>('Validators')
+    },
+    isFunction(api.query.nominationPools?.minCreateBond) && {
+      name: 'pools',
+      text: t<string>('Pools')
     }
   ]);
 
@@ -51,7 +72,10 @@ function StakingApp ({ basePath }: Props): React.ReactElement<Props> {
         items={itemsRef.current}
       />
       <Switch>
-        <Route>
+        <Route path={pathRef.current.pools}>
+          <Pools />
+        </Route>
+        <Route path={basePath}>
           <Validators
             favorites={favorites}
             isRelay={isRelay}
